pyproject.toml
[project]
name = "pdf-extraction-demo"
version = "0.1.0"
description = "Local bake-off for column-aware PDF extraction with pdfplumber and PDF Oxide"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "markitdown[pdf]>=0.1.3",
  "pdfplumber>=0.11.0",
  "pdf-oxide>=0.3.14",
  "pymupdf (>=1.27.2,<2.0.0)",
  "camelot-py[cv]>=0.11.0",
  "psutil>=6.0.0",
  "pypdf>=4.0.0",
  "docling>=2.0.0",
  "marker-pdf>=1.0.0",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
